refactor(radio): move Radio to its folder with theme interface
#443 #234 #458
